Moved adapter files into folder
And also cleaned things up a bit
