+++ api/ExpressedRealms.Email/EmailClientAdapter/IEmailClientAdapter.cs
@@ -1,0 +1,6 @@
+namespace ExpressedRealms.Email.EmailClientAdapter;
+
+internal interface IEmailClientAdapter
+{
+    Task SendEmailAsync(EmailData data);
+}
