--- api/ExpressedRealms.Email/IEmailClientAdapter.cs
@@ -1,6 +1,0 @@
-namespace ExpressedRealms.Email;
-
-public interface IEmailClientAdapter
-{
-    Task SendEmailAsync(EmailData data);
-}
